NFC: split out stress test to it's own file
